+++ manul/src/protocol/object_safe.rs
@@ -1,197 +1,40 @@
-use alloc::{boxed::Box, collections::BTreeMap, format};
-use core::{fmt::Debug, marker::PhantomData};
-
-use rand_core::CryptoRngCore;
+use alloc::{boxed::Box, format};
 
 use super::{
-    errors::{LocalError, ReceiveError},
-    message::{DirectMessage, EchoBroadcast, NormalBroadcast, ProtocolMessage},
-    round::{Artifact, CommunicationInfo, FinalizeOutcome, PartyId, Payload, Protocol, Round},
-    round_id::{RoundId, TransitionInfo},
-    serialization::{Deserializer, Serializer},
+    errors::LocalError,
+    round::{PartyId, Protocol, Round},
+    round_id::RoundId,
 };
 
-// Since we want `Round` methods to take `&mut impl CryptoRngCore` arguments
-// (which is what all cryptographic libraries generally take), it cannot be object-safe.
-// Thus we have to add this crate-private object-safe layer on top of `Round`.
-pub(crate) trait ObjectSafeRound<Id: PartyId>: 'static + Debug + Send + Sync {
-    type Protocol: Protocol<Id>;
-
-    fn transition_info(&self) -> TransitionInfo;
-
-    fn communication_info(&self) -> CommunicationInfo<Id>;
-
-    fn make_direct_message(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        deserializer: &Deserializer,
-        destination: &Id,
-    ) -> Result<(DirectMessage, Option<Artifact>), LocalError>;
-
-    fn make_echo_broadcast(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        deserializer: &Deserializer,
-    ) -> Result<EchoBroadcast, LocalError>;
-
-    fn make_normal_broadcast(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        deserializer: &Deserializer,
-    ) -> Result<NormalBroadcast, LocalError>;
-
-    fn receive_message(
-        &self,
-        deserializer: &Deserializer,
-        from: &Id,
-        message: ProtocolMessage,
-    ) -> Result<Payload, ReceiveError<Id, Self::Protocol>>;
-
-    fn finalize(
-        self: Box<Self>,
-        rng: &mut dyn CryptoRngCore,
-        payloads: BTreeMap<Id, Payload>,
-        artifacts: BTreeMap<Id, Artifact>,
-    ) -> Result<FinalizeOutcome<Id, Self::Protocol>, LocalError>;
-
-    /// Returns the type ID of the implementing type.
-    fn get_type_id(&self) -> core::any::TypeId {
-        core::any::TypeId::of::<Self>()
-    }
-}
-
-// The `fn(Id)` bit is so that `ObjectSafeRoundWrapper` didn't require a bound on `Id` to be `Send + Sync`.
-#[derive(Debug)]
-pub(crate) struct ObjectSafeRoundWrapper<Id, R> {
-    round: R,
-    phantom: PhantomData<fn(Id)>,
-}
-
-impl<Id, R> ObjectSafeRoundWrapper<Id, R>
-where
-    Id: PartyId,
-    R: Round<Id>,
-{
-    pub fn new(round: R) -> Self {
-        Self {
-            round,
-            phantom: PhantomData,
-        }
-    }
-}
-
-impl<Id, R> ObjectSafeRound<Id> for ObjectSafeRoundWrapper<Id, R>
-where
-    Id: PartyId,
-    R: Round<Id>,
-{
-    type Protocol = <R as Round<Id>>::Protocol;
-
-    fn transition_info(&self) -> TransitionInfo {
-        self.round.transition_info()
-    }
-
-    fn communication_info(&self) -> CommunicationInfo<Id> {
-        self.round.communication_info()
-    }
-
-    fn make_direct_message(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        #[allow(unused_variables)] deserializer: &Deserializer,
-        destination: &Id,
-    ) -> Result<(DirectMessage, Option<Artifact>), LocalError> {
-        self.round.make_direct_message(rng, serializer, destination)
-    }
-
-    fn make_echo_broadcast(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        #[allow(unused_variables)] deserializer: &Deserializer,
-    ) -> Result<EchoBroadcast, LocalError> {
-        self.round.make_echo_broadcast(rng, serializer)
-    }
-
-    fn make_normal_broadcast(
-        &self,
-        rng: &mut dyn CryptoRngCore,
-        serializer: &Serializer,
-        #[allow(unused_variables)] deserializer: &Deserializer,
-    ) -> Result<NormalBroadcast, LocalError> {
-        self.round.make_normal_broadcast(rng, serializer)
-    }
-
-    fn receive_message(
-        &self,
-        deserializer: &Deserializer,
-        from: &Id,
-        message: ProtocolMessage,
-    ) -> Result<Payload, ReceiveError<Id, Self::Protocol>> {
-        self.round.receive_message(deserializer, from, message)
-    }
-
-    fn finalize(
-        self: Box<Self>,
-        rng: &mut dyn CryptoRngCore,
-        payloads: BTreeMap<Id, Payload>,
-        artifacts: BTreeMap<Id, Artifact>,
-    ) -> Result<FinalizeOutcome<Id, Self::Protocol>, LocalError> {
-        self.round.finalize(rng, payloads, artifacts)
-    }
-}
-
-// We do not want to expose `ObjectSafeRound` to the user, so it is hidden in a struct.
 /// A wrapped new round that may be returned by [`Round::finalize`]
 /// or [`EntryPoint::make_round`](`crate::protocol::EntryPoint::make_round`).
 #[derive_where::derive_where(Debug)]
-pub struct BoxedRound<Id: PartyId, P: Protocol<Id>> {
-    wrapped: bool,
-    round: Box<dyn ObjectSafeRound<Id, Protocol = P>>,
-}
+pub struct BoxedRound<Id: PartyId, P: Protocol<Id>>(Box<dyn Round<Id, Protocol = P>>);
 
 impl<Id: PartyId, P: Protocol<Id>> BoxedRound<Id, P> {
     /// Wraps an object implementing the dynamic round trait ([`Round`](`crate::protocol::Round`)).
     pub fn new_dynamic<R: Round<Id, Protocol = P>>(round: R) -> Self {
-        Self {
-            wrapped: true,
-            round: Box::new(ObjectSafeRoundWrapper::new(round)),
-        }
+        Self(Box::new(round))
     }
 
-    pub(crate) fn new_object_safe<R: ObjectSafeRound<Id, Protocol = P>>(round: R) -> Self {
-        Self {
-            wrapped: false,
-            round: Box::new(round),
-        }
+    pub(crate) fn as_ref(&self) -> &dyn Round<Id, Protocol = P> {
+        self.0.as_ref()
     }
 
-    pub(crate) fn as_ref(&self) -> &dyn ObjectSafeRound<Id, Protocol = P> {
-        self.round.as_ref()
-    }
-
-    pub(crate) fn into_boxed(self) -> Box<dyn ObjectSafeRound<Id, Protocol = P>> {
-        self.round
+    pub(crate) fn into_boxed(self) -> Box<dyn Round<Id, Protocol = P>> {
+        self.0
     }
 
     fn boxed_type_is<T: 'static>(&self) -> bool {
-        core::any::TypeId::of::<T>() == self.round.get_type_id()
+        core::any::TypeId::of::<T>() == self.0.get_type_id()
     }
 
     /// Attempts to extract an object of a concrete type, preserving the original on failure.
     pub fn try_downcast<T: Round<Id>>(self) -> Result<T, Self> {
-        if self.wrapped && self.boxed_type_is::<ObjectSafeRoundWrapper<Id, T>>() {
+        if self.boxed_type_is::<T>() {
             // Safety: This is safe since we just checked that we are casting to the correct type.
-            let boxed_downcast = unsafe {
-                Box::<ObjectSafeRoundWrapper<Id, T>>::from_raw(
-                    Box::into_raw(self.round) as *mut ObjectSafeRoundWrapper<Id, T>
-                )
-            };
-            Ok(boxed_downcast.round)
+            let boxed_downcast = unsafe { Box::<T>::from_raw(Box::into_raw(self.0) as *mut T) };
+            Ok(*boxed_downcast)
         } else {
             Err(self)
         }
@@ -209,8 +52,8 @@ impl<Id: PartyId, P: Protocol<Id>> BoxedRound<Id, P> {
     ///
     /// Fails if the wrapped type is not `T`.
     pub fn downcast_ref<T: Round<Id>>(&self) -> Result<&T, LocalError> {
-        if self.wrapped && self.boxed_type_is::<ObjectSafeRoundWrapper<Id, T>>() {
-            let ptr: *const dyn ObjectSafeRound<Id, Protocol = P> = self.round.as_ref();
+        if self.boxed_type_is::<T>() {
+            let ptr: *const dyn Round<Id, Protocol = P> = self.0.as_ref();
             // Safety: This is safe since we just checked that we are casting to the correct type.
             Ok(unsafe { &*(ptr as *const T) })
         } else {
@@ -226,6 +69,6 @@ impl<Id: PartyId, P: Protocol<Id>> BoxedRound<Id, P> {
         // This constructs a new `TransitionInfo` object, so calling this method inside `Session`
         // has mild performance drawbacks.
         // This is mostly exposed for the sake of users writing `Misbehave` impls for testing.
-        self.round.transition_info().id()
+        self.0.transition_info().id()
     }
 }